Rename to @anio-software/pkg-private.node-fs-find-nearest-file-factory:
Ok so the issue is that if we are using top level await to load node dependencies then this code will be included in every package that imports this code.

Even if the import isn't used, the code still ends up being in the bundle.

Not good. As a reminder, this function needs to be:

1) Runnable in the browser without throwing an error (because of joytest)
2) Needs to have a synchronous API (also because of joytest)

The compromise is going to be a factory function (that is async) that returns a sync function that later can be used.

I.e. if the factory function isn't called, the code won't end up in the output bundle.